Fix Earth's heightmap and remove artificial noise
Set the deformity to 9350 and the offset to -1147. Those values correctly set
the altitude of Bonneville, Salar Uyuni, and Etosha Pan. Those locations were
chosen because they have very different altitudes and they are flat. The
flatness guarantees that there are no aliasing problems with the heightmap.

Because the deformity was reduced, the noise needed to be zeroed to prevent
random parts of Florida from being underwater and random islands popping up.

Fixes KSP-RO/RealSolarSystem#149